Added Encoder i2c read and ROS msg publish
Added TEENSY_ENCODER_ADDR 0x30 - address setup in Teensy as i2c Address
Added EncRecor struct, a struct of 2 int16's to match what's being sent from the encoder. Honestly, this could be increased to a larger integer to not roll over so fast, anything the teensy could store. That would be up to int32. But then I'm not sure how many bytes to read in the i2c read. Anyway, that's a small implementation update

Setup the ROS note to publis pub_enc as a int16 multi array

Added i2c read to read at TEENSY_ENCODER_ADDR. This required opening a file for ioctl. I think that's how it's done.

Finally, added a pub.publish for pub_enc,